Keep Host upgrades compatible with attachments
