+++ linux-console/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aru-host",
-  "version": "0.30.0",
+  "version": "0.30.1",
   "private": true,
   "type": "module",
   "license": "Apache-2.0",
