+++ tests/macos-installer-smoke.sh
@@ -18,7 +18,6 @@ node --check "$SELFHOST_DIR/mobile-collaborator-replicas.mjs"
 node --check "$SELFHOST_DIR/collaborator-surfaces.mjs"
 node --check "$SELFHOST_DIR/collaborator-surface-bundles.mjs"
 node --check "$SELFHOST_DIR/collaborator-conversations.mjs"
-node --check "$SELFHOST_DIR/collaborator-conversation-attachments.mjs"
 node --check "$SELFHOST_DIR/collaborator-initiative.mjs"
 node --check "$SELFHOST_DIR/collaborator-projects.mjs"
 node --check "$SELFHOST_DIR/apns-push.mjs"
@@ -53,7 +52,6 @@ test -f "$alpha/current/mobile-collaborator-replicas.mjs"
 test -f "$alpha/current/collaborator-surfaces.mjs"
 test -f "$alpha/current/collaborator-surface-bundles.mjs"
 test -f "$alpha/current/collaborator-conversations.mjs"
-test -f "$alpha/current/collaborator-conversation-attachments.mjs"
 test -f "$alpha/current/collaborator-initiative.mjs"
 test -f "$alpha/current/collaborator-projects.mjs"
 test -f "$alpha/current/apns-push.mjs"
@@ -123,7 +121,7 @@ test -x "$beta/current/server.mjs"
 bash "$SELFHOST_DIR/package-macos-release.sh" "$artifacts/aru-selfhost-macos.tar.gz" 0.28.0-test >/dev/null
 test -s "$artifacts/aru-selfhost-macos.tar.gz"
 test -s "$artifacts/aru-selfhost-macos.tar.gz.sha256"
-for file in aru-selfhost-stub.mjs backup-settings.mjs conversation-turn-relay.mjs collaborator-host.mjs mobile-collaborator-replicas.mjs collaborator-cognition.mjs collaborator-surfaces.mjs collaborator-surface-bundles.mjs collaborator-conversations.mjs collaborator-conversation-attachments.mjs collaborator-initiative.mjs collaborator-projects.mjs apns-push.mjs codex-app-server-driver.mjs direct-api-driver.mjs provider-profiles.mjs provider-secret-store.mjs node-control.mjs node-workspaces.mjs plugin-supervisor.mjs plugin-workshop.mjs source-plugin-runtime.mjs source-plugin-runner.mjs run-node.sh install-macos.sh aru-selfhostctl-macos; do
+for file in aru-selfhost-stub.mjs backup-settings.mjs conversation-turn-relay.mjs collaborator-host.mjs mobile-collaborator-replicas.mjs collaborator-cognition.mjs collaborator-surfaces.mjs collaborator-surface-bundles.mjs collaborator-conversations.mjs collaborator-initiative.mjs collaborator-projects.mjs apns-push.mjs codex-app-server-driver.mjs direct-api-driver.mjs provider-profiles.mjs provider-secret-store.mjs node-control.mjs node-workspaces.mjs plugin-supervisor.mjs plugin-workshop.mjs source-plugin-runtime.mjs source-plugin-runner.mjs run-node.sh install-macos.sh aru-selfhostctl-macos; do
   tar -tzf "$artifacts/aru-selfhost-macos.tar.gz" | grep -Fqx "$file"
 done
 tar -xOf "$artifacts/aru-selfhost-macos.tar.gz" release.json \
